refactor: rename NavbarDropdownItem -> NavbarMenuItem
