+++ public/assets/images/icons/NavbarMenuIcons.tsx
