+++ src/features/common/Layout/Navbar/NavbarMenu.module.scss
@@ -2,7 +2,7 @@
 
 // new navbar dropdown styles
 
-.navbarDropdownItem {
+.navbarMenuItem {
   display: flex;
   align-items: center;
   gap: 10px;
